requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI Integration
anthropic==0.77.1
openai==2.16.0
httpx==0.28.1

# Database & Caching
sqlalchemy==2.0.25
aiosqlite==0.19.0
redis==5.0.1
python-dotenv==1.0.0

# Task Queue
celery==5.3.4
flower==2.0.1
kombu==5.3.4

# WebSocket & Real-time
websockets==12.0
python-socketio==5.11.0

# Data Processing
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3

# Logging & Monitoring
structlog==24.1.0
rich==13.7.0

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Additional Required Packages
requests==2.32.5
Pillow==11.3.0
PyPDF2==3.0.1
reportlab==4.4.9
sexpdata==1.0.2